feat: initial targets
@alessandrotenaglia @lorenzo-bianchi @giorgiomanca97
Please have a look at the `docker/container-x86-*/commands.sh`: there is a special command that needs to be MANUALLY issued when testing with your PC directly connected to the robot via Eth, specifying the name of your NICs.
Again, this is for testing, for when you connect your PC to the robot with a cable, configuration on Jetson boards having both WiFi and Eth NICs is automatic.
It is required to configure network interfaces for Cyclone DDS.
If you want to know what needs to be done to use the robot on a different SBC than its own NX, have a look at one of the `docker/container-*/Dockerfile`; keep in mind that `x86-*` and `jetson5*` ones differ because the latter already comes with rmw_cyclonedds_cpp installed.
Also, first project based on DUA 2.0.